Submitting two solutions for the findPairs problem.
The first solution is a 'bruteforce' approach that uses two nested loops to find the sum of each element with the next ones in the array to find the pairs that add up to the target.

The second solution uses a hash set, which makes it more efficient to find a solution because we only need to loop once.

One thing to consider is that the first solution cuts the array short because it only considers positive values. If negative values were in the equation, then the array should not be filtered.